+++ cmd/cmd.go
@@ -14,7 +14,6 @@ import (
 	"time"
 
 	"github.com/micro/micro/v3/client/cli/util"
-	mucmd "github.com/micro/micro/v3/internal/command"
 	uconf "github.com/micro/micro/v3/internal/config"
 	"github.com/micro/micro/v3/internal/helper"
 	"github.com/micro/micro/v3/internal/network"
@@ -250,12 +249,12 @@ func action(c *cli.Context) error {
 		// execute the Config.Set RPC, setting the flags in the
 		// request.
 		if srv, ns, err := lookupService(c); err != nil {
-			return mucmd.CliError(err)
+			return util.CliError(err)
 		} else if srv != nil && shouldRenderHelp(c) {
 			return cli.Exit(formatServiceUsage(srv, c), 1)
 		} else if srv != nil {
 			if err := callService(srv, ns, c); err != nil {
-				return mucmd.CliError(err)
+				return util.CliError(err)
 			}
 			return nil
 		}
